+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Before releasing:
 ### Added
 
 - Implemented a conversion between `bool` and `LogicLevel`. (#428)
+- Added support for capturing backtraces without heap allocations via the `BacktraceIter` struct. (#456)
 
 ### Fixed
 
@@ -47,10 +48,13 @@ Before releasing:
 ### Changed
 
 - Corrected some minor doc comments and updated the Nix flake. (#451)
-
 - Programs now execute in Thumb mode instead of ARM mode, reducing code size. (#454)
+- Backtraces are now captured using the frame pointer instead of through unwind tables. (#456) (**Breaking Change**)
 
 ### Removed
+
+- The vexide linker script now removes unwind tables from the executable. (#456) (**Breaking Change**)
+- The `backtrace` Cargo feature has been removed; the `vexide::backtrace` module is now available unconditionally. (#456) (**Breaking Change**)
 
 ### New Contributors